Rename package pitop.virtual_hardware to pitop.simulation
