+++ packages/simulation/README.rst
@@ -1,0 +1,6 @@
+===========================
+pi-top Python Simulation
+===========================
+
+Creates graphical simulations of pi-top hardware, either reflecting real
+harware or fully virtual.
