+++ packages/simulation/pitop/simulation/images/LED_red_on.png
