+++ packages/simulation/pitop/simulation/images/LED_yellow_off.png
