+++ packages/simulation/pitop/simulation/images/LED_yellow_on.png
