--- packages/virtual_hardware/MANIFEST.in
@@ -1,1 +1,0 @@
-recursive-include pitop/virtual_hardware/ *
