--- packages/virtual_hardware/README.rst
@@ -1,5 +1,0 @@
-==============================
-pi-top Python Virtual Hardware
-==============================
-
-Tools for visualising and mocking pi-top hardware.
