+++ tests/test_pitop.py
@@ -140,7 +140,7 @@ def test_blockpi_rover(rover):
 
 def test_pitop_simulate(pitop, mocker, create_sim, snapshot):
     mocker.patch(
-        "pitop.virtual_hardware.simulation.sprites.is_virtual_hardware",
+        "pitop.simulation.sprites.is_virtual_hardware",
         return_value=True,
     )
 
@@ -175,7 +175,7 @@ def test_pitop_simulate(pitop, mocker, create_sim, snapshot):
 def test_pitop_visualize(pitop, create_sim, mocker, snapshot):
     # with is_virtual_hardware False, pygame button events will not be handled
     mocker.patch(
-        "pitop.virtual_hardware.simulation.sprites.is_virtual_hardware",
+        "pitop.simulation.sprites.is_virtual_hardware",
         return_value=False,
     )
